requirements.txt
python-dateutil>=2.8.2
python-dotenv>=0.20.0
tweepy==3.10.0
twitchAPI==2.3.2
